Extract migration parsing from VM::AR::Controller to a concern
